+++ apps/mcp-server/RELEASES.md
@@ -31,3 +31,11 @@ Based on:
 - Speakeasy CLI 1.771.0 (2.893.0) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [mcp-typescript v0.1.0] apps/mcp-server
+
+## 2026-06-12 00:27:40
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.777.1 (2.903.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [mcp-typescript v0.2.0] apps/mcp-server